requirements.txt
traci
tensorflow
ultralytics
paho-mqtt
opencv-python
